+++ SignNow.Net/Model/Responses/FolderIdentityResponse.cs
@@ -1,0 +1,10 @@
+namespace SignNow.Net.Model.Responses
+{
+    public class FolderIdentityResponse
+    {
+        /// <summary>
+        /// Identity of then newly created/renamed folder.
+        /// </summary>
+        public string Id { get; set; }
+    }
+}